+++ backend/app/Http/Controllers/LessonController.php
@@ -3,13 +3,28 @@
 namespace App\Http\Controllers;
 
 use App\Models\Lesson;
+use App\Models\UserHasSubject;
 use Illuminate\Http\Request;
 
 class LessonController extends Controller
 {
-    public function index()
+    public function index(Request $request )
     {
-        return Lesson::all();
+        $id_class=$request->input('id_class');
+        $id_subject=$request->input('id_subject');
+
+        if(!$id_class && !$id_subject){
+            return Lesson::all(); }
+        if(!$id_class && $id_subject)
+        {$id_user_class_class = \DB::table('lesson')->where('subject_id_subject','=',$id_subject)->get();
+            return $id_user_class_class;}
+        if($id_class && !$id_subject)
+        {$id_user_class_class = \DB::table('lesson')->where('class_id_class','=',$id_class)->get();
+            return $id_user_class_class;}
+        if($id_class && $id_subject)
+        {$id_user_class_class = \DB::table('lesson')->where([['class_id_class','=',$id_class],['subject_id_subject','=',$id_subject]])->get();
+            return $id_user_class_class;}
+
     }
 
     public function show(Lesson $lesson)
@@ -18,24 +33,37 @@ class LessonController extends Controller
     }
 
     public function store(Request $request)
-    {
-        $lesson = Lesson::create($request->all());
+    {$isteacher=$request->input('isteacher');
+        $date=$request->input('date');
+        $class_id_class=$request->input('Class_id_class');
+        $subject_id_subject=$request->input('Subject_id_subject');
 
-        return response()->json($lesson, 201);
+
+        if($isteacher==3){
+            $lesson = Lesson::create(['date'=> $date,'Class_id_class'=> $class_id_class,'Subject_id_subject'=> $subject_id_subject]);
+
+            return response()->json($lesson, 201);}
     }
 
     public function update(Request $request, Lesson $lesson)
-    {
-        $lesson->update($request->all());
+    {$isteacher=$request->input('isteacher');
+        $date=$request->input('date');
+        $class_id_class=$request->input('Class_id_class');
+        $subject_id_subject=$request->input('Subject_id_subject');
 
-        return response()->json($lesson, 200);
+
+        if($isteacher==3){
+            $lesson->update(['date'=> $date,'Class_id_class'=> $class_id_class,'Subject_id_subject'=> $subject_id_subject]);
+
+            return response()->json($lesson, 200);}
     }
 
-    public function delete(Lesson $lesson)
-    {
-        $lesson->delete();
+    public function delete(Request $request, Lesson $lesson)
+    {$isteacher=$request->input('isteacher');
+        if($isteacher==3){
+            $lesson->delete();
 
-        return response()->json(null, 204);
+            return response()->json(null, 204);}
     }
