+++ backend/app/Http/Controllers/SubjectController.php
@@ -18,24 +18,37 @@ class SubjectController extends Controller
     }
 
     public function store(Request $request)
-    {
-        $subject = Subject::create($request->all());
+    {$isteacher=$request->input('isteacher');
+        $subject_name=$request->input('subject_name');
+        if( $isteacher == 3){
+            $id_user_class = \DB::table('subject')->where('Subject_name','=',$subject_name)->get();
 
-        return response()->json($subject, 201);
+            if (!count($id_user_class)){
+                $subject = Subject::create(['subject_name'=> $subject_name]);
+
+                return response()->json($subject, 201);}}
     }
 
     public function update(Request $request, Subject $subject)
-    {
-        $subject->update($request->all());
+    {$isteacher=$request->input('isteacher');
+        $subject_name=$request->input('subject_name');
+        if( $isteacher == 3){
+            $id_user_class = \DB::table('subject')->where('Subject_name','=',$subject_name)->get();
 
-        return response()->json($subject, 200);
+            if (!count($id_user_class)){
+                $subject ->update(['subject_name'=> $subject_name]);
+                return response()->json($subject, 200);}}
+
+
+
+
     }
 
-    public function delete(Subject $subject)
-    {
-        $subject->delete();
+    public function delete(Request $request, Subject $subject)
+    {$isteacher=$request->input('isteacher');
 
-        return response()->json(null, 204);
+        if( $isteacher == 3){$subject->delete();
+            return response()->json(null, 204);}
     }
